Moved ColorMap mixin to series folder and removed namespace.
